+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/craft_student/craft/open_weater_app/src/App.jsx":"1","/home/craft_student/craft/open_weater_app/src/index.js":"2","/home/craft_student/craft/open_weater_app/src/reportWebVitals.js":"3","/home/craft_student/craft/open_weater_app/src/components/Showcaser.jsx":"4","/home/craft_student/craft/open_weater_app/src/components/DailyForecast.jsx":"5"},{"size":3317,"mtime":1612690849470,"results":"6","hashOfConfig":"7"},{"size":287,"mtime":1612367729778,"results":"8","hashOfConfig":"7"},{"size":362,"mtime":1612347649397,"results":"9","hashOfConfig":"7"},{"size":1534,"mtime":1612632532917,"results":"10","hashOfConfig":"7"},{"size":1320,"mtime":1612633208472,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ecdbx6",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"19"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/craft_student/craft/open_weater_app/src/App.jsx",[],"/home/craft_student/craft/open_weater_app/src/index.js",[],["24","25"],"/home/craft_student/craft/open_weater_app/src/reportWebVitals.js",[],["26","27"],"/home/craft_student/craft/open_weater_app/src/components/Showcaser.jsx",[],"/home/craft_student/craft/open_weater_app/src/components/DailyForecast.jsx",[],{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","replacedBy":"31"},{"ruleId":"28","replacedBy":"32"},{"ruleId":"30","replacedBy":"33"},"no-native-reassign",["34"],"no-negated-in-lhs",["35"],["34"],["35"],"no-global-assign","no-unsafe-negation"]
+[{"/home/craft_student/craft/open_weater_app/src/App.jsx":"1","/home/craft_student/craft/open_weater_app/src/index.js":"2","/home/craft_student/craft/open_weater_app/src/reportWebVitals.js":"3","/home/craft_student/craft/open_weater_app/src/components/Showcaser.jsx":"4","/home/craft_student/craft/open_weater_app/src/components/DailyForecast.jsx":"5","/home/craft_student/craft/open_weater_app/src/components/TempChart.jsx":"6"},{"size":2417,"mtime":1612691257371,"results":"7","hashOfConfig":"8"},{"size":287,"mtime":1612367729778,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1612347649397,"results":"10","hashOfConfig":"8"},{"size":1534,"mtime":1612632532917,"results":"11","hashOfConfig":"8"},{"size":1320,"mtime":1612633208472,"results":"12","hashOfConfig":"8"},{"size":1089,"mtime":1612691132295,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ecdbx6",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"21"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/craft_student/craft/open_weater_app/src/App.jsx",[],"/home/craft_student/craft/open_weater_app/src/index.js",[],["28","29"],"/home/craft_student/craft/open_weater_app/src/reportWebVitals.js",[],["30","31"],"/home/craft_student/craft/open_weater_app/src/components/Showcaser.jsx",[],"/home/craft_student/craft/open_weater_app/src/components/DailyForecast.jsx",[],"/home/craft_student/craft/open_weater_app/src/components/TempChart.jsx",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"32","replacedBy":"36"},{"ruleId":"34","replacedBy":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],["38"],["39"],"no-global-assign","no-unsafe-negation"]
