+++ src/App.jsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import axios from 'axios'
 import { Header, Tab } from 'semantic-ui-react'
-import { Line } from 'react-chartjs-2'
+
 import Showcaser from './components/Showcaser'
 import DailyForecast from './components/DailyForecast'
+import TempChart from './components/TempChart'
 
 class App extends React.Component {
   state = {
@@ -39,12 +40,7 @@ class App extends React.Component {
       {
         menuItem: 'Daily Temp',
         render: () => <Tab.Pane attached={false}>
-          {dailyTemp && <Line
-            data={data}
-            width={500}
-            height={250}
-            options={{ maintainAspectRatio: false }}
-          />}
+          <TempChart dailyTemp={dailyTemp} />
         </Tab.Pane>,
       },
       {
@@ -55,44 +51,11 @@ class App extends React.Component {
       },
     ]
 
-    let labels = []
-    let dataItems = []
-    let dataItems2 = []
-    let data
-    if (dailyTemp) {
-      dailyTemp.forEach(day => {
-        labels.push(new Date(day.dt * 1000).toLocaleDateString('sv'))
-        dataItems.push(day.temp.day)
-        dataItems2.push(day.feels_like.day)
-      })
-      data = {
-        labels: labels,
-        datasets: [
-          {
-            label: 'Daily Temperature',
-            data: dataItems,
-            borderColor: '#B55DFF',
-            backgroundColor: '#32C5CE',
-          },
-          {
-            label: 'Daily Feels Like',
-            data: dataItems2,
-            borderColor: '#B55DFF',
-            backgroundColor: '#5883AB',
-          }
-        ]
-      }
-    }
-    // new Date(dailyTemp[0].dt * 1000).toLocaleDateString('sv')
-
     return (
       <div className="main-container" data-cy="weather-display">
         <Header color="yellow" size="huge" textAlign="center">Your Weather Forecast</Header>
-
         <Showcaser weatherInfo={weatherInfo}>
-
           <Tab menu={{ pointing: true }} panes={panes} />
-
         </Showcaser>
 
       </div >
